StorageAccountBlobsIngestion/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
requests
azure-storage-file-share==12.3.0
azure-storage-blob
